move _default_per_datum_log_likelihood to core/protocols.py
Completes the likelihood-protocol relocation: the length-1-batch fallback now
lives next to ConditionallyIndependentLikelihood, the protocol it serves, and
the protocol docstring names it explicitly again. modeling/_likelihood.py
keeps only the conditioning machinery; the sole importer (the
conditionally-independent-likelihood tests) is repointed. Private helper, so
no back-compat re-export.
